moved queue component in backend to game/queue
