+++ backend/src/game/game.module.ts
@@ -1,7 +1,7 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { GameService } from './game.service';
 import { GameGateway } from './game.gateway';
-import { QueueModule } from '../queue/queue.module';
+import { QueueModule } from './queue/queue.module';
 import { SequelizeModule } from '@nestjs/sequelize';
 import { Match } from './entities/match.entity';
